Add repeated answers to the log as well.
If the user answers with the same draggable-droppable combination
again, it is now saved in the log like normal answers, but with
a new additional key rerun (value true). If the user clicks
on answered droppables to review their feedback, the clicks are
also stored in the log now. The click events have an additional
key "click" with value true. Clicks are only stored while
the exercise is unfinished as the log event is sent to the server
the moment the exercise is finished.